creating tweeter API
I have managed to use my credentials from dev tweeter to extract some data. I have saved two files, one with the credentials save as a json file , and the other is the basic code to extract the tweeter data which still needs to be edited to extract the desired tweets.